requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
plotly>=5.18.0
matplotlib>=3.7.0
streamlit>=1.29.0
streamlit-folium>=0.15.0
folium>=0.15.0
python-dateutil>=2.8.0
pyyaml>=6.0
tqdm>=4.66.0